+++ app/uses/page.tsx
@@ -1,13 +1,8 @@
-'use client'
-
 import * as React from 'react'
 
-import * as TabsPrimitive from '@radix-ui/react-tabs'
-import { motion } from 'framer-motion'
-import Link from 'next/link'
+import Tabs from './components/Tabs'
 
 import { MotionHeader, MotionMain } from '~components/ContentWrappers'
-import { List, ListItem } from '~components/List'
 import Prose from '~components/Prose'
 import SectionTitle from '~components/SectionTitle'
 
@@ -15,152 +10,7 @@ export const metadata = {
   title: 'Uses',
 }
 
-function Software() {
-  return (
-    <List>
-      <ListItem
-        icon="vs-code"
-        label="VS Code"
-        href="https://code.visualstudio.com/"
-      >
-        I've tried them all; Dreamweaver, Sublime Text, BB Edit, Textmate, Atom,
-        etc.
-      </ListItem>
-      <ListItem icon="figma" label="Figma" href="https://www.figma.com">
-        Previously Sketch, but Figma is just too good to ignore; with a plethora
-        of community-driven plugins.
-      </ListItem>
-      <ListItem
-        icon="rectangle"
-        label="Rectangle"
-        href="https://rectangleapp.com/"
-      >
-        The easiest window-management tool I've ever used.
-      </ListItem>
-      <ListItem
-        icon="chrome"
-        label="Google Chrome"
-        href="https://www.google.com/chrome"
-      >
-        More recently Firefox, but Chrome's consistency keeps bringing me back.
-      </ListItem>
-      <ListItem icon="terminal" label="Warp" href="https://www.warp.dev/">
-        The built-in terminal in VS Code is fine; though I'm fortunate enough to
-        have the screen real-estate for a dedicated, always-open terminal.
-      </ListItem>
-      <ListItem
-        icon="github"
-        label="GitHub Desktop"
-        href="https://desktop.github.com/"
-      >
-        I do feel more accomplished in the terminal at-times, but the interface
-        and ease-of-use can't be understated for all my projects.
-      </ListItem>
-      <ListItem
-        icon="mimestream"
-        label="Mimestream"
-        href="https://mimestream.com/"
-      >
-        Handles multiple Google-based email accounts like a champ.
-      </ListItem>
-      <ListItem icon="1password" label="1Password" href="https://1password.com">
-        Synced across all my devices; never had an issue.
-      </ListItem>
-      <ListItem icon="reminders" label="Apple Reminders">
-        I've tried a few different apps, but the iCloud integration and
-        feature-set has always been enough for me.
-      </ListItem>
-    </List>
-  )
-}
-
-function Gear() {
-  return (
-    <List>
-      <ListItem
-        href="https://amzn.to/3AuEVVV"
-        icon="mbpro"
-        label="MacBook Pro (15-inch, 2019)"
-      >
-        Stuck with the 'ole touch-bar for now, but very much looking forward to
-        giving the M2 models a shot.
-      </ListItem>
-      <ListItem
-        href="https://amzn.to/3QXI333"
-        icon="iphone"
-        label="iPhone 13 Pro"
-      >
-        I swap for a new model every 2-3 years.
-      </ListItem>
-      <ListItem
-        href="https://amzn.to/3ABcerh"
-        icon="watch"
-        label="Apple Watch Series 7 45mm"
-      >
-        Longtime Fitbit user, but notifications work so much better now. I
-        thought I'd miss the 4-5 days of battery life, but I get a solid 2 days.
-      </ListItem>
-      <ListItem icon="monitor" label="Monitors">
-        <Link href="https://amzn.to/3K5ZSe9">LG 32UL750-W</Link> front and
-        center, with a{' '}
-        <Link href="https://amzn.to/3pxBokr">Dell UltraSharp U2720Q</Link>{' '}
-        rotated 90º to my left.
-      </ListItem>
-      <ListItem icon="mouse" label="Logitech MX Masters">
-        Took almost no time to get used to either the{' '}
-        <Link href="https://amzn.to/3PCyVzG">keyboard</Link> or{' '}
-        <Link href="https://amzn.to/3ccyLS8">mouse</Link>.
-      </ListItem>
-      <ListItem
-        href="https://www.aersf.com/travel-pack-3-black"
-        icon="backpack"
-        label="Aer Travel Pack 3"
-      >
-        I spent way too much time researching backpacks, but this is a
-        much-needed, dual-purpose upgrade that I'm happy to add.
-      </ListItem>
-      <ListItem
-        href="https://opalcamera.com/"
-        icon="webcam"
-        label="Opal Camera"
-      >
-        I'll eventually switch over to Continuity in iOS 16, but this has worked
-        well so far. I've tried{' '}
-        <Link href="https://amzn.to/3QEn99r">Lumina</Link> and{' '}
-        <Link href="https://amzn.to/3T1SQel">Logitech Brio</Link>.
-      </ListItem>
-      <ListItem
-        href="https://amzn.to/3TnaoSf"
-        icon="dock"
-        label="CalDigit TS3 Plus Dock"
-      >
-        Having a single cord (with power) that connects all my devices to my
-        laptop is great!
-      </ListItem>
-      <ListItem
-        href="https://amzn.to/3dGpYIp"
-        icon="stand"
-        label="Rain Design mStand"
-      >
-        Had this since 2016 and still looks/works great.
-      </ListItem>
-    </List>
-  )
-}
-
-const tabs: { title: string; value: string }[] = [
-  {
-    title: 'Software',
-    value: 'software',
-  },
-  {
-    title: 'Gear',
-    value: 'gear',
-  },
-]
-
 export default function Uses() {
-  const [tab, setTab] = React.useState('software')
   return (
     <>
       <MotionHeader>
@@ -175,35 +25,7 @@ export default function Uses() {
             </p>
           </Prose>
         </div>
-        <TabsPrimitive.Root value={tab} onValueChange={setTab}>
-          <TabsPrimitive.TabsList
-            aria-label="What I use"
-            className="mb-8 flex items-center gap-6"
-          >
-            {tabs.map(({ title, value }) => (
-              <React.Fragment key={value}>
-                <TabsPrimitive.TabsTrigger
-                  className="relative text-lg font-semibold text-slate-11 transition-colors hocus:text-slate-12 rdx-state-active:text-slate-12"
-                  value={value}
-                >
-                  {title}
-                  {value === tab && (
-                    <motion.span
-                      className="absolute inset-x-0 -bottom-1.5 h-0.5 rounded-full bg-gradient-to-r from-primary-9 to-accent-9"
-                      layoutId="underline"
-                    />
-                  )}
-                </TabsPrimitive.TabsTrigger>
-              </React.Fragment>
-            ))}
-          </TabsPrimitive.TabsList>
-          <TabsPrimitive.Content value="software">
-            <Software />
-          </TabsPrimitive.Content>
-          <TabsPrimitive.Content value="gear">
-            <Gear />
-          </TabsPrimitive.Content>
-        </TabsPrimitive.Root>
+        <Tabs />
       </MotionMain>
     </>
   )
